requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
langchain==0.1.17
langchain-groq==0.1.0
pydantic==1.10.11
jinja2==3.1.2
numpy==1.25.2
